added common transform util object
