+++ src/main/scala/com/hhi/sap/analysis/ANL_THD_MRPL_WEEK.scala
@@ -1,6 +1,7 @@
 package com.hhi.sap.analysis
 
 import com.hhi.sap.analysis.functions.MRPLTableUtils
+import com.hhi.sap.analysis.functions.common.TransformUtils
 import org.apache.spark.sql.{DataFrame, SQLContext}
 import org.slf4j.LoggerFactory
 
@@ -16,10 +17,10 @@ class ANL_THD_MRPL_WEEK(sql: SQLContext) {
     val underRDD = MRPLTableUtils.getWeekTable(mrplRDD.filter(_.week <= -5), -5)
     val upperRDD = MRPLTableUtils.getWeekTable(mrplRDD.filter(_.week >= 20), 20)
 
-    MRPLTableUtils
+    TransformUtils
       .makeUnion(mrplRDD.filter(-4 until 19 contains _.week).toDF(), underRDD, upperRDD)
-      .transform(MRPLTableUtils.addSERNO)
-      .transform(MRPLTableUtils.pivotTable)
-      .transform(MRPLTableUtils.mappingMRPLTable)
+      .transform(TransformUtils.addSERNO)
+      .transform(TransformUtils.pivotTableByCount)
+      .transform(TransformUtils.mappingTable)
   }
 }
