+++ src/main/scala/com/hhi/sap/analysis/ANL_THD_WEIGHT_WEEK.scala
@@ -1,6 +1,7 @@
 package com.hhi.sap.analysis
 
 import com.hhi.sap.analysis.functions.WeightTableUtils
+import com.hhi.sap.analysis.functions.common.TransformUtils
 import org.apache.spark.sql.{DataFrame, SQLContext}
 import org.slf4j.LoggerFactory
 
@@ -17,10 +18,10 @@ class ANL_THD_WEIGHT_WEEK(sql: SQLContext) {
     val underRDD = WeightTableUtils.getWeightTable(weightRDD.filter(_.week <= -5), -5)
     val upperRDD = WeightTableUtils.getWeightTable(weightRDD.filter(_.week >= 20), 20)
 
-    WeightTableUtils
+    TransformUtils
       .makeUnion(weightRDD.filter(-4 until 19 contains _.week).toDF(), underRDD, upperRDD)
-      .transform(WeightTableUtils.addSERNO)
-      .transform(WeightTableUtils.pivotTable)
-      .transform(WeightTableUtils.mappingMRPLTable)
+      .transform(TransformUtils.addSERNO)
+      .transform(TransformUtils.pivotTableByMenge)
+      .transform(TransformUtils.mappingTable)
   }
 }
